:recycle: refactor: #S10P21C101-backend 책 평점 수정

+++ backend/src/main/java/com/ssafy/backend/domain/book/entity/BookReview.java
@@ -29,7 +29,7 @@ public class BookReview {
     private String content;
 
     @Builder
-    public BookReview(User user, Book book, int score, String content) {
+    public BookReview(User user, Book book, double score, String content) {
         this.user = user;
         this.book = book;
         this.score = score;
